refactor: 챗봇 레이어드 아키텍처 구조에 맞게 수정
