+++ backend/src/main/java/com/ice/studyroom/domain/chatbot/infrastructure/persistence/ChatbotCategoryRepository.java
@@ -1,5 +1,6 @@
-package com.ice.studyroom.domain.chatbot.domain.category;
+package com.ice.studyroom.domain.chatbot.infrastructure.persistence;
 
+import com.ice.studyroom.domain.chatbot.domain.entity.ChatbotCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatbotCategoryRepository extends JpaRepository<ChatbotCategory, String> {
