+++ backend/src/main/java/com/ice/studyroom/domain/chatbot/infrastructure/persistence/ChatbotQuestionRepository.java
@@ -1,7 +1,8 @@
-package com.ice.studyroom.domain.chatbot.domain.question;
+package com.ice.studyroom.domain.chatbot.infrastructure.persistence;
 
 import java.util.List;
 
+import com.ice.studyroom.domain.chatbot.domain.entity.ChatbotQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
